fix: diffusion head 强制 float32 (fp16 softmax 溢出)

+++ src/vla/model_loader.py
@@ -148,7 +148,7 @@ class OpenVLAPolicy:
 
         self.diffusion_head = DiffusionActionHead(
             hidden_dim=hidden_dim, chunk_size=self.chunk_size,
-        ).to(device=self.device, dtype=self.llm.dtype)
+        ).to(device=self.device, dtype=torch.float32)
 
         import torchvision.transforms as T
         self.img_preprocess = T.Compose([
